+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Change Log
 
-## 1.1.6 (2023-02-)
+## 1.1.6 (2023-02-28)
 
 - `isValidDate`: Only the `yyyy-mm-dd` format can be verified
 - `dateToYYYYMMDD`: Add `dateToYYYYMMDD` method
 - `createDateListFromRange`: Add `createDateListFromRange` method
+- `arrCount`: Add `arrCount` method
 
 ## 1.1.5 (2023-02-07)